refactor: rename schema sorting DSL helpers and move PostgreSQL strategy
Apply review feedback from #2459: prefix core sorting DSL helpers with schema_ (schema_sort_by_name/type/type_and_name/metadata), rename pgsql_sort_strategy to pgsql_schema_sort_by_type, and move the PostgreSQL strategy to Flow\ETL\Adapter\PostgreSql\Schema\SortingStrategy\TypeStrategy.
